requirements.txt
fastapi
uvicorn
pydantic
google-genai
ollama